LPS-91967 Move reusable logic into FileUtil
