+++ modules/sdk/gradle-plugins-node/src/main/java/com/liferay/gradle/plugins/node/tasks/NpmInstallTask.java
@@ -26,14 +26,10 @@ import java.io.File;
 import java.io.IOException;
 
 import java.nio.charset.StandardCharsets;
-import java.nio.file.DirectoryStream;
-import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.SimpleFileVisitor;
 import java.nio.file.StandardCopyOption;
-import java.nio.file.attribute.BasicFileAttributes;
 
 import java.util.Arrays;
 import java.util.List;
@@ -290,65 +286,6 @@ public class NpmInstallTask extends ExecuteNpmTask {
 		return completeArgs;
 	}
 
-	private static void _createBinDirLinks(Logger logger, File nodeModulesDir)
-		throws IOException {
-
-		JsonSlurper jsonSlurper = new JsonSlurper();
-
-		Path nodeModulesDirPath = nodeModulesDir.toPath();
-
-		Path nodeModulesBinDirPath = nodeModulesDirPath.resolve(
-			_NODE_MODULES_BIN_DIR_NAME);
-
-		try (DirectoryStream<Path> directoryStream = Files.newDirectoryStream(
-				nodeModulesDirPath, _directoryStreamFilter)) {
-
-			for (Path dirPath : directoryStream) {
-				Path packageJsonPath = dirPath.resolve("package.json");
-
-				if (Files.notExists(packageJsonPath)) {
-					continue;
-				}
-
-				Map<String, Object> packageJsonMap =
-					(Map<String, Object>)jsonSlurper.parse(
-						packageJsonPath.toFile());
-
-				Object binObject = packageJsonMap.get("bin");
-
-				if (!(binObject instanceof Map<?, ?>)) {
-					continue;
-				}
-
-				Map<String, String> binJsonMap = (Map<String, String>)binObject;
-
-				if (binJsonMap.isEmpty()) {
-					continue;
-				}
-
-				Files.createDirectories(nodeModulesBinDirPath);
-
-				for (Map.Entry<String, String> entry : binJsonMap.entrySet()) {
-					String linkFileName = entry.getKey();
-					String linkTargetFileName = entry.getValue();
-
-					Path linkPath = nodeModulesBinDirPath.resolve(linkFileName);
-					Path linkTargetPath = dirPath.resolve(linkTargetFileName);
-
-					Files.deleteIfExists(linkPath);
-
-					Files.createSymbolicLink(linkPath, linkTargetPath);
-
-					if (logger.isInfoEnabled()) {
-						logger.info(
-							"Created binary symbolic link {} which targets {}",
-							linkPath, linkTargetPath);
-					}
-				}
-			}
-		}
-	}
-
 	private static String _getNodeModulesCacheDigest(
 		NpmInstallTask npmInstallTask) {
 
@@ -419,12 +356,12 @@ public class NpmInstallTask extends ExecuteNpmTask {
 			FileUtil.syncDir(
 				project, nodeModulesCacheDir, nodeModulesDir, nativeSync);
 
-			_removeBinDirLinks(logger, nodeModulesDir);
+			FileUtil.removeBinDirLinks(logger, nodeModulesDir);
 		}
 		else {
 			npmInstallTask._npmInstall(reset);
 
-			_removeBinDirLinks(logger, nodeModulesDir);
+			FileUtil.removeBinDirLinks(logger, nodeModulesDir);
 
 			if (logger.isLifecycleEnabled()) {
 				logger.lifecycle(
@@ -437,41 +374,8 @@ public class NpmInstallTask extends ExecuteNpmTask {
 		}
 
 		if (!OSDetector.isWindows()) {
-			_createBinDirLinks(logger, nodeModulesDir);
+			FileUtil.createBinDirLinks(logger, nodeModulesDir);
 		}
-	}
-
-	private static void _removeBinDirLinks(
-			final Logger logger, File nodeModulesDir)
-		throws IOException {
-
-		Files.walkFileTree(
-			nodeModulesDir.toPath(),
-			new SimpleFileVisitor<Path>() {
-
-				@Override
-				public FileVisitResult preVisitDirectory(
-						Path dirPath, BasicFileAttributes basicFileAttributes)
-					throws IOException {
-
-					String dirName = String.valueOf(dirPath.getFileName());
-
-					if (dirName.equals(_NODE_MODULES_BIN_DIR_NAME)) {
-						if (logger.isInfoEnabled()) {
-							logger.info(
-								"Removing binary symbolic links from {}",
-								dirPath);
-						}
-
-						FileUtil.deleteSymbolicLinks(dirPath);
-
-						return FileVisitResult.SKIP_SUBTREE;
-					}
-
-					return FileVisitResult.CONTINUE;
-				}
-
-			});
 	}
 
 	private File _getExistentFile(String fileName) {
@@ -594,18 +498,6 @@ public class NpmInstallTask extends ExecuteNpmTask {
 
 		Files.write(shrinkwrapJsonPath, json.getBytes(StandardCharsets.UTF_8));
 	}
-
-	private static final String _NODE_MODULES_BIN_DIR_NAME = ".bin";
-
-	private static final DirectoryStream.Filter<Path> _directoryStreamFilter =
-		new DirectoryStream.Filter<Path>() {
-
-			@Override
-			public boolean accept(Path path) throws IOException {
-				return Files.isDirectory(path);
-			}
-
-		};
 
 	private Object _nodeModulesCacheDir;
 	private boolean _nodeModulesCacheNativeSync = true;
